pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mitre-attackctl"
version = "0.1.1"
description = "A fast, batteries-included CLI companion for MITRE ATT&CK® TTPs"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Jonathan Haas", email = "jonathan@haasonsaas.com"}
]
maintainers = [
    {name = "Jonathan Haas", email = "jonathan@haasonsaas.com"}
]
keywords = ["mitre", "attack", "ttps", "cybersecurity", "cli", "threat-hunting", "security", "blue-team"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "rapidfuzz>=3.0.0",
    "requests>=2.31.0",
    "stix2>=3.0.0",
    "taxii2-client>=2.3.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
ai = [
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/haasonsaas/mitre-attackctl"
Repository = "https://github.com/haasonsaas/mitre-attackctl"
Issues = "https://github.com/haasonsaas/mitre-attackctl/issues"
Documentation = "https://github.com/haasonsaas/mitre-attackctl#readme"
Changelog = "https://github.com/haasonsaas/mitre-attackctl/blob/main/CHANGELOG.md"

[project.scripts]
attackctl = "attackctl.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"